+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Improve `unsquashfs` and `add` cli args to match `squashfs-tools/unsquashfs` cli
 - `Filesystem::push_file(..)` now takes for bytes anything that is `into Read` instead of `into Vec<u8>`
+- `Filesystem::from_reader(..)`, `R` now takes `Read + Seek` instead our own `ReadSeek`
+- `Filesystem::from_reader_with_offset(..)`, `R` now takes `Read + Seek` instead our own `ReadSeek`
 
 ## [v0.6.0] - 2023-01-10
 - Fix bug in our filesystem tree causing directory header information (gui, uid, permissions)